requirements.txt
# Core dependencies
anthropic>=0.39.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Arcade (for animation rendering)
arcade==3.3.3